Jason Duell fix:
The default names for the inverse database files
created with the -q option violate the regular cscope
naming convention (they get called cscope.in.out and
cscope.po.out, whereas if you call "cscope -f cscope",
they get called cscope.out.in and cscope.out.po). This
causes cscope to fail if you create the database
without "-f cscope" but then try to load it with "-f
cscope". Since the vim editor plugin for cscope always
uses the -f flag, this makes it impossible to use the
vim plugin unless you use the -f flag to create the
database (and the editor freezes when you try to load
the database). This is a needless hassle for new users.

While the easiest fix for this problem would be to
change to default names for the -q files to match the
regular naming convention, this could break user setups
that rely on the existing names. So this patch takes a
different tack. Whenever "cscope.out" is loaded as the
main database, both "cscope.po.out" and "cscope.out.po"
are checked (ditto for cscope.in.out & cscope.out.in)
before. The two combinations are checked in the
appropriate order (i.e., "cscope.po.out" is checked
first if the -f flag was not used, and "cscope.out.po"
if it was). Only if neither file is present does the
database load fail.

The patch also contains a rewrite of the documentation
for the -q option. I remember being confused by it
when I first read the docs, and it now also contains
information about the -f flag that is superfluous.